move useWindowOverflow to hooks
