+++ src/components/ColorPicker.tsx
@@ -1,9 +1,10 @@
 import { rgbToHex } from '@mui/material'
-import React, { FC, useLayoutEffect, useRef, useState } from 'react'
+import React, { FC, useRef } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { formatSelectionActionCreator as formatSelection } from '../actions/formatSelection'
 import { isTouch } from '../browser'
 import * as selection from '../device/selection'
+import useWindowOverflow from '../hooks/useWindowOverflow'
 import getThoughtById from '../selectors/getThoughtById'
 import themeColors from '../selectors/themeColors'
 import commandStateStore from '../stores/commandStateStore'
@@ -14,26 +15,6 @@ import TextColorIcon from './icons/TextColor'
 
 /** A function that adds an alpha channel to a hex color. */
 const addAlphaToHex = (hex: string) => (hex.length === 7 ? hex + 'ff' : hex)
-
-/** A hook that returns the left and right overflow of the element outside the bounds of the screen. Do not re-calculate on every render or it will create an infinite loop when scrolling the Toolbar. */
-const useWindowOverflow = (ref: React.RefObject<HTMLElement>) => {
-  const [overflow, setOverflow] = useState({ left: 0, right: 0 })
-
-  useLayoutEffect(() => {
-    if (!ref.current) return
-    const rect = ref.current.getBoundingClientRect()
-    // Subtract the previous overflow, since that affects the client rect.
-    // Otherwise the overflow will alternate on each render as it moves on and off the screen.
-    const left = Math.max(0, -rect.x + 15 - overflow.left)
-    // add 10px for padding
-    const right = Math.max(0, rect.x + rect.width - window.innerWidth + 10 - overflow.right)
-    if (left > 0 || right > 0) {
-      setOverflow({ left, right })
-    }
-  }, [ref, overflow.left, overflow.right])
-
-  return overflow
-}
 
 /** A small, square color swatch that can be picked in the color picker. */
 const ColorSwatch: FC<{
